feat(database-types): use extracted types
This adds the `enableDatabaseExtractionScan` configuration.

When enabled, it will gather the column type information from the
`database/schema.php` file. Or the files defined in `databaseExtractionSchemaPath`.

Those files will be generated using our extraction tool and should be commited to the repository.

With this feature, we can provide type information for columns that are based
on more complex or raw SQL statements.

It overrides the table definitions read from the squashed migration helper with actual types from the database.
It still allows to gather further info from migrations that alter the schema.

It also solves these problems:
- Squashed Postgres schemas can be correctly inferred (Larastan otherwise uses a Mysql-based SQL parser, which does not work reliably with PostgreSQL syntax / types)
- Columns created by schema macros can be correctly inferred